Link to our new API documentation
As the old site, developer.twingly.com, has been shut down. The documentation for
all of our APIs are now available (behind login) at https://app.twingly.com instead.

I didn't link to the specific documentation sections in the code, just so we don't
need to update the links in case we restructure the documentation in the future.